A few lines on using the docker container
I tried getting this to work on Windows but fell down a rabbit hole chasing the correct IP number to use. But the docker method worked perfectly the first time, which was great. It connected to R just fine, and I could do many of the tasks in the vignettes, very fun!